requirements.txt
# Core ML Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Gradient Boosting
lightgbm>=4.0.0
xgboost>=2.0.0
catboost>=1.2.0

# Explainability
shap>=0.43.0

# Deep Learning (optional)
# torch>=2.0.0
# transformers>=4.30.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Data Validation
great-expectations>=0.17.0
pandera>=0.17.0

# Feature Store
feast>=0.32.0

# Experiment Tracking
mlflow>=2.8.0

# Monitoring
evidently>=0.4.0
prometheus-client>=0.17.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0
